% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate stations}
\usage{
aggregate(stations_df, tag)
}
\arguments{
\item{stations_df}{data.frame containing stations to aggregate}

\item{tag}{string with the tag consulted.}
}
\value{
a data.frame with only two columns containing the dates and the
aggregated observations.
}
\description{
Aggregate stations observations.
}
\keyword{internal}
