% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_annual}
\alias{aggregate_annual}
\title{Calculate annual aggregate of climate data}
\usage{
aggregate_annual(monthly_data, FUN)
}
\arguments{
\item{monthly_data}{\code{data.frame} with monthly aggregated data}

\item{FUN}{Function to use for aggregation}
}
\value{
\code{data.frame} with annual aggregated data
}
\description{
Calculate annual aggregate of climate data
}
\keyword{internal}
