% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_bshg}
\alias{aggregate_bshg}
\title{Aggregate sunshine duration data for different frequencies}
\usage{
aggregate_bshg(.data, frequency)
}
\arguments{
\item{.data}{\code{data.frame} obtained from download functions}

\item{frequency}{character with the aggregation frequency. (\code{"day"},
\code{"month"} or \code{"year"})}
}
\value{
\code{data.frame} with the aggregated data
}
\description{
Calculate the aggregate of downloaded sunshine duration data to day, month
and year
}
\examples{
\dontrun{
aggregate_bshg(.data, "day")
}

}
