% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_daily}
\alias{aggregate_daily}
\title{Calculate daily aggregate of climate data}
\usage{
aggregate_daily(hourly_data, FUN)
}
\arguments{
\item{hourly_data}{\code{data.frame} with hourly aggregated data}

\item{FUN}{Function to use for aggregation}
}
\value{
\code{data.frame} with daily aggregated data
}
\description{
Calculate daily aggregate of climate data
}
\keyword{internal}
