% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_monthly}
\alias{aggregate_monthly}
\title{Calculate monthly aggregate of climate data}
\usage{
aggregate_monthly(daily_data, FUN)
}
\arguments{
\item{daily_data}{\code{data.frame} with daily aggregated data}

\item{FUN}{Function to use for aggregation}
}
\value{
\code{data.frame} object with monthly aggregated data
}
\description{
Calculate monthly aggregate of climate data
}
\keyword{internal}
