% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_ptpm}
\alias{aggregate_ptpm}
\title{Aggregate precipitation data for different frequencies}
\usage{
aggregate_ptpm(.data, frequency)
}
\arguments{
\item{.data}{\code{data.frame} obtained from download functions}

\item{frequency}{character with the aggregation frequency. (\code{"day"},
\code{"month"} or \code{"year"})}
}
\value{
\code{data.frame} with the aggregated data
}
\description{
Calculate the aggregate of downloaded precipitation data to day, month
and year
}
\examples{
\dontrun{
aggregate_ptpm(.data, "day")
}

}
