% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_tmx}
\alias{aggregate_tmx}
\title{Aggregate maximum temperature data for different frequencies}
\usage{
aggregate_tmx(.data, frequency)
}
\arguments{
\item{.data}{\code{data.frame} obtained from download functions}

\item{frequency}{character with the aggregation frequency. (
\code{"month"} or \code{"year"})}
}
\value{
\code{data.frame} with the aggregated data
}
\description{
Calculate the maximum of downloaded maximum temperature data to day, month
and year
}
\examples{
\dontrun{
aggregate_tmx(.data, "day")
}

}
