% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_tssm}
\alias{aggregate_tssm}
\title{Aggregate dry-bulb temperature data for different frequencies}
\usage{
aggregate_tssm(.data, frequency)
}
\arguments{
\item{.data}{\code{data.frame} obtained from download functions}

\item{frequency}{character with the aggregation frequency. (\code{"day"},
\code{"month"} or \code{"year"})}
}
\value{
\code{data.frame} with the aggregated data
}
\description{
Calculate the mean of downloaded hourly dry-bulb temperature data to day,
month and year
}
\examples{
\dontrun{
aggregate_tssm(.data, "day")
}

}
