% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{annual_bshg}
\alias{annual_bshg}
\title{Calculate annual sunshine duration}
\usage{
annual_bshg(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the year}
}
\value{
numeric value calculated
}
\description{
Calculate annual sunshine duration
}
\keyword{internal}
