% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{annual_ptpm}
\alias{annual_ptpm}
\title{Calculate annual precipitation}
\usage{
annual_ptpm(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the year}
}
\value{
numeric value calculated
}
\description{
Calculate annual precipitation
}
\keyword{internal}
