% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{annual_tmn}
\alias{annual_tmn}
\title{Calculate annual minimum temperature}
\usage{
annual_tmn(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the year}
}
\value{
numeric value calculated
}
\description{
Calculate annual minimum temperature
}
\keyword{internal}
