% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{annual_tmx}
\alias{annual_tmx}
\title{Calculate annual maximum temperature}
\usage{
annual_tmx(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the year}
}
\value{
numeric value calculated
}
\description{
Calculate annual maximum temperature
}
\keyword{internal}
