% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{annual_tssm}
\alias{annual_tssm}
\title{Calculate annual dry-bulb mean temperature}
\usage{
annual_tssm(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the year}
}
\value{
numeric value calculated
}
\description{
Calculate annual dry-bulb mean temperature
}
\keyword{internal}
