% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{clean_ddp_pp}
\alias{clean_ddp_pp}
\title{Clean and pivot population projection data}
\usage{
clean_ddp_pp(dataset)
}
\arguments{
\item{dataset}{data.frame of population projections to subset}
}
\value{
data.frame of pivoted and cleaned population projection data
}
\description{
This function provides the needed structure to pivot population projection
from the original format, making easier to access and subset.
}
\examples{
\dontrun{
clean_dpd_pp
}
}
\keyword{internal}
