% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{code_to_name_dep}
\alias{code_to_name_dep}
\title{Retrieve departments' DIVIPOLA names from codes}
\usage{
code_to_name_dep(department_code)
}
\arguments{
\item{department_code}{character vector with the DIVIPOLA codes of the
departments.}
}
\value{
character vector with the DIVIPOLA name of the departments.
}
\description{
Retrieve departments' DIVIPOLA official names from their DIVIPOLA codes.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dptos <- c("73", "05", "11")
code_to_name_dep(dptos)
\dontshow{\}) # examplesIf}
}
