% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{code_to_name_mun}
\alias{code_to_name_mun}
\title{Retrieve municipalities' DIVIPOLA names from codes}
\usage{
code_to_name_mun(municipality_code)
}
\arguments{
\item{municipality_code}{character vector with the DIVIPOLA codes of the
municipalities.}
}
\value{
character vector with the DIVIPOLA name of the municipalities.
}
\description{
Retrieve municipalities' DIVIPOLA official names from their DIVIPOLA codes.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mpios <- c("73001", "11001", "05615")
code_to_name_mun(mpios)
\dontshow{\}) # examplesIf}
}
