% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{codes_to_string}
\alias{codes_to_string}
\title{Turn numeric vector of municipality codes into character vector}
\usage{
codes_to_string(codes)
}
\arguments{
\item{codes}{numeric vector with municipality codes. They need to include the
department and the municipality code.}
}
\value{
character vector with municipality codes in standardized code
}
\description{
Turn numeric vector of municipality codes into character vector
}
\keyword{internal}
