% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{daily_bshg}
\alias{daily_bshg}
\title{Calculate daily sunshine duration}
\usage{
daily_bshg(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the day}
}
\value{
numeric value calculated
}
\description{
Calculate daily sunshine duration
}
\keyword{internal}
