% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{daily_tssm}
\alias{daily_tssm}
\title{Calculate daily dry-bulb mean temperature}
\usage{
daily_tssm(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the day}
}
\value{
numeric value calculated
}
\description{
Calculate daily dry-bulb mean temperature
}
\keyword{internal}
