% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Download data dictionary}
\usage{
data_dictionary(dataset)
}
\arguments{
\item{dataset}{String indicating dataset code.}
}
\value{
tibble containing data dictionary
}
\description{
Download data dictionary
}
\examples{
data_dictionary("DANE_MGNCNPV_2018_SETU")

}
