% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_2018.R
\docType{data}
\name{dengue_2018}
\alias{dengue_2018}
\title{Dengue data 2018}
\format{
\subsection{\code{dengue_2018}}{

data frame with 33 rows and 2 columns:
\describe{
\item{DPTO_CCDGO}{Department code}
\item{CASES}{Dengue case count}
}
}
}
\source{
\url{https://epiverse-trace.github.io/sivirep/}
}
\usage{
dengue_2018
}
\description{
Count of Dengue cases in Colombia for 2021 grouped at department level
}
\keyword{datasets}
