% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_DPTO.R
\docType{data}
\name{DICT_DANE_MGN_2018_DPTO}
\alias{DICT_DANE_MGN_2018_DPTO}
\alias{dpto}
\title{DICT_DANE_MGN_2018_DPTO}
\format{
An object of class \code{data.frame} with 86 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_DPTO)
}
\description{
dictionary of variables presented in the geospatial departmental
dataset
}
\details{
Dictionary for geospatial departmental dataset
}
\keyword{datasets}
