% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_MPIO.R
\docType{data}
\name{DICT_DANE_MGN_2018_MPIO}
\alias{DICT_DANE_MGN_2018_MPIO}
\alias{mpio}
\title{DICT_DANE_MGN_2018_MPIO}
\format{
An object of class \code{data.frame} with 88 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_MPIO)
}
\description{
dictionary of variables presented in the geospatial
municipality dataset
}
\details{
Dictionary for geospatial municipality dataset
}
\keyword{datasets}
