% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_MPIOCL.R
\docType{data}
\name{DICT_DANE_MGN_2018_MPIOCL}
\alias{DICT_DANE_MGN_2018_MPIOCL}
\alias{mpiocl}
\title{DICT_DANE_MGN_2018_MPIOCL}
\format{
An object of class \code{data.frame} with 89 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_MPIOCL)
}
\description{
dictionary of variables presented in the geospatial municipality
(including class) dataset
}
\details{
Dictionary for geospatial municipality (including class) dataset
}
\keyword{datasets}
