% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_MZN.R
\docType{data}
\name{DICT_DANE_MGN_2018_MZN}
\alias{DICT_DANE_MGN_2018_MZN}
\alias{mzn}
\title{DICT_DANE_MGN_2018_MZN}
\format{
An object of class \code{data.frame} with 106 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_MZN)
}
\description{
dictionary of variables presented in the geospatial block
dataset
}
\details{
Dictionary for geospatial block dataset
}
\keyword{datasets}
