% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_SECR.R
\docType{data}
\name{DICT_DANE_MGN_2018_SECR}
\alias{DICT_DANE_MGN_2018_SECR}
\alias{secr}
\title{DICT_DANE_MGN_2018_SECR}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_SECR)
}
\description{
dictionary of variables presented in the geospatial rural sector
dataset
}
\details{
Dictionary for geospatial rural sector dataset
}
\keyword{datasets}
