% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_SECU.R
\docType{data}
\name{DICT_DANE_MGN_2018_SECU}
\alias{DICT_DANE_MGN_2018_SECU}
\alias{secu}
\title{DICT_DANE_MGN_2018_SECU}
\format{
An object of class \code{data.frame} with 99 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_SECU)
}
\description{
dictionary of variables presented in the geospatial urban sector
dataset
}
\details{
Dictionary for geospatial urban sector dataset
}
\keyword{datasets}
