% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_SETR.R
\docType{data}
\name{DICT_DANE_MGN_2018_SETR}
\alias{DICT_DANE_MGN_2018_SETR}
\alias{setr}
\title{DICT_DANE_MGN_2018_SETR}
\format{
An object of class \code{data.frame} with 99 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_SETR)
}
\description{
dictionary of variables presented in the geospatial rural
section dataset
}
\details{
Dictionary for geospatial rural section dataset
}
\keyword{datasets}
