% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_SETU.R
\docType{data}
\name{DICT_DANE_MGN_2018_SETU}
\alias{DICT_DANE_MGN_2018_SETU}
\alias{setu}
\title{DICT_DANE_MGN_2018_SETU}
\format{
An object of class \code{data.frame} with 97 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_SETU)
}
\description{
dictionary of variables presented in the geospatial urban
section dataset
}
\details{
Dictionary for geospatial urban section dataset
}
\keyword{datasets}
