% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DICT_DANE_MGN_2018_ZU.R
\docType{data}
\name{DICT_DANE_MGN_2018_ZU}
\alias{DICT_DANE_MGN_2018_ZU}
\alias{zu}
\title{DICT_DANE_MGN_2018_ZU}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\usage{
data(DICT_DANE_MGN_2018_ZU)
}
\description{
dictionary of variables presented in the geospatial urban zone
dataset
}
\details{
Dictionary for geospatial urban zone dataset
}
\keyword{datasets}
