% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionary_geospatial}
\alias{dictionary_geospatial}
\title{Download data dictionary for geospatial dataframes}
\usage{
dictionary_geospatial(dataset)
}
\arguments{
\item{dataset}{String indicating dataset code.}
}
\value{
tibble containing data dictionary
}
\description{
Download data dictionary for geospatial dataframes
}
\examples{
data_dictionary("DANE_MGNCNPV_2018_SETU")

}
