% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{divipola_department_code}
\alias{divipola_department_code}
\title{Retrieve departments' DIVIPOLA codes from names}
\usage{
divipola_department_code(department_name)
}
\arguments{
\item{department_name}{character vector with the names of the departments}
}
\value{
character vector with the DIVIPOLA codes of the departments
}
\description{
Retrieve departments' DIVIPOLA codes from their names
}
\examples{
code_tolima <- divipola_department_code("TOLIMA")

}
