% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{divipola_department_name}
\alias{divipola_department_name}
\title{Retrieve departments' DIVIPOLA names from codes}
\usage{
divipola_department_name(department_code)
}
\arguments{
\item{department_name}{character vector with the DIVIPOLA codes of the
departments}
}
\value{
character vector with the DIVIPOLA name of the departments
}
\description{
Retrieve departments' DIVIPOLA official names from their DIVIPOLA codes
}
\examples{
name_tolima <- divipola_department_name("73")

}
