% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{divipola_table}
\alias{divipola_table}
\title{Retrieve DIVIPOLA table}
\usage{
divipola_table()
}
\value{
\code{data.frame} object with DIVIPOLA table
}
\description{
Retrieve DIVIPOLA table including departments and municipalities
}
\examples{
divipola <- divipola_table()

}
