% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download dataset}
\usage{
download(dataset)
}
\arguments{
\item{dataset}{dataset code}
}
\value{
dataframe downloaded
}
\description{
Download dataset
}
\examples{
download("MGNCNPV01")

}
