% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate_data}
\alias{download_climate_data}
\title{Download climate data from stations}
\usage{
download_climate_data(
  stations,
  start_date,
  end_date,
  frequency,
  tags,
  aggregate = TRUE
)
}
\arguments{
\item{stations}{Vector containing the stations' codes.}

\item{start_date}{Character with the first date to consult in format
"YYYY-MM-DD".}

\item{end_date}{Character with the last date to consult in format
"YYYY-MM-DD". Last available date is 2023/05/31.}

\item{frequency}{Character with the aggregation frequency. Can be
"day", "week","month" or "year".}

\item{tags}{Character containing tags to consult.}

\item{aggregate}{Boolean for data aggregation, if FALSE, returns the data from
each individual station in ROI.(Default = TRUE).}
}
\value{
data.frame with the observed data for the given geometry.
}
\description{
Retrieve climate stations with observed data in a given geometry. This
data is retrieved from local meteorological stations.
}
\examples{
\dontrun{
climate_stations(stations, "2021-11-14", "2021-11-30", "day", "TSSM_CON")
}

}
