% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate_geom}
\alias{download_climate_geom}
\title{Download climate data from geometry}
\usage{
download_climate_geom(
  geometry,
  start_date,
  end_date,
  frequency,
  tags,
  aggregate = TRUE
)
}
\arguments{
\item{geometry}{\code{sf} geometry containing the geometry for a given ROI.
This geometry can be either a \code{POLYGON} or \code{MULTIPOLYGON}}

\item{start_date}{character with the first date to consult in the format
\code{"YYYY-MM-DD"}}

\item{end_date}{character with the last date to consult in the format
\code{"YYYY-MM-DD"} (Last available date is \code{"2023-05-31"})}

\item{frequency}{character with the aggregation frequency. (\code{"day"},
\code{"week"}, \code{"month"} or \code{"year"})}

\item{tags}{character containing climate tags to consult}

\item{aggregate}{logical for data aggregation, if \code{FALSE}, returns the
data from each individual station in the area. (Default = \code{TRUE})}
}
\value{
\code{data.frame} with observations from the stations in the area
}
\description{
Download climate data from stations contained in a geometry. This data is
retrieved from local meteorological stations provided by IDEAM.
}
\examples{
lat <- c(5.166278, 5.166278, 4.982247, 4.982247, 5.166278)
lon <- c(-75.678072, -75.327859, -75.327859, -75.678072, -75.678072)
polygon <- sf::st_polygon(x = list(cbind(lon, lat)))
geometry <- sf::st_sfc(polygon)
roi <- sf::st_as_sf(geometry)
tssm <- download_climate_geom(roi, "2021-11-14", "2021-11-20",
"day", "TSSM_CON")
print(tssm)

}
