% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate_mpio}
\alias{download_climate_mpio}
\title{Retrieve climate data by municipality name}
\usage{
download_climate_mpio(
  name,
  start_date,
  end_date,
  frequency,
  tags,
  group = FALSE
)
}
\arguments{
\item{name}{Character with the DIVIPOLA code for the municipality.}

\item{start_date}{Character with the first date to consult in format
"YYYY-MM_DD".}

\item{end_date}{Character with the last date to consult in format
"YYYY-MM_DD".}

\item{frequency}{Character with the aggregation frequency. Can be
"day", "week","month" or"year".}

\item{tags}{Character containing tags to consult.}

\item{group}{Boolean for data aggregation, if FALSE, returns the data from
each individual station in ROI.(Default = FALSE).}
}
\value{
data.frame with the observed data for the given municipality
}
\description{
Retrieve climate stations with observed data in a given municipality.
}
\examples{
\dontrun{
download_climate_mpio("11001", "2021-11-14", "2021-11-30", "day", "TSSM_CON")
}

}
