% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate_stations}
\alias{download_climate_stations}
\title{Download climate data from stations}
\usage{
download_climate_stations(
  stations,
  start_date,
  end_date,
  frequency,
  tags,
  aggregate = TRUE
)
}
\arguments{
\item{stations}{numeric vector containing the stations' codes}

\item{start_date}{character with the first date to consult in the format
\code{"YYYY-MM-DD"}}

\item{end_date}{character with the last date to consult in the format
\code{"YYYY-MM-DD"} (Last available date is \code{"2023-05-31"})}

\item{frequency}{character with the aggregation frequency. (\code{"day"},
\code{"week"}, \code{"month"} or \code{"year"})}

\item{tags}{character containing climate tags to consult}

\item{aggregate}{logical for data aggregation, if \code{FALSE}, returns the
data from each individual station in the area. (Default = \code{TRUE})}
}
\value{
\code{data.frame} with observations from the requested stations
}
\description{
Download climate data from named stations.This data is
retrieved from local meteorological stations provided by IDEAM.
}
\examples{
stations <- c(26155110, 26155170)
tssm <- download_climate_stations(
  stations, "2021-11-14", "2021-11-20",
  "day", "TSSM_CON", TRUE
)
print(tssm)

}
