% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_demographic.R
\name{download_demographic}
\alias{download_demographic}
\title{Download demographic dataset}
\usage{
download_demographic(dataset)
}
\arguments{
\item{dataset}{character with the demographic dataset name. Please use
\code{list_datasets("demographic", "EN")} or
\code{list_datasets("demographic", "ES")} to check available datasets.}
}
\value{
\code{data.frame} object with downloaded data.
}
\description{
This function downloads demographic datasets from the National Population
and Dwelling Census (CNPV) of 2018.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
house_under_15 <- download_demographic("DANE_CNPVH_2018_1HD")
head(house_under_15)
\dontshow{\}) # examplesIf}
}
