% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_mgn_cnpv}
\alias{filter_mgn_cnpv}
\title{Filter MGN data}
\usage{
filter_mgn_cnpv(.data, ad = NULL, columns = NULL, include_geometry = TRUE)
}
\arguments{
\item{.data}{dataframe containing MGN - CNPV data}

\item{ad}{character vector with municipalities to filter}

\item{columns}{character vector with columns to filter}

\item{include_geometry}{boolean to include (or not) the geometry.}
}
\value{
sf dataframe containing MGN(National Geo statistical Frame) and CNPV
(National Population and Living Census) data
}
\description{
Filter MGN data
}
