% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_mgn_cnpv_ad2}
\alias{filter_mgn_cnpv_ad2}
\title{Filter municipal dataset}
\usage{
filter_mgn_cnpv_ad2(census_data, codes, columns)
}
\arguments{
\item{census_data}{MGN - CNPV dataset}

\item{codes}{Administrative division to include}

\item{columns}{Columns to include}
}
\value{
filtered dataset for municipal level
}
\description{
Filter municipal dataset
}
\keyword{internal}
