% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_dictionaries.R
\docType{data}
\name{geospatial_dictionaries}
\alias{geospatial_dictionaries}
\title{geospatial_dictionaries}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(geospatial_dictionaries)
}
\description{
dictionaries of variables presented in geospatial datasets
}
\details{
Dictionaries for geospatial datasets in english and spanish
}
\keyword{datasets}
