% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{get_climate_tags}
\alias{get_climate_tags}
\title{List climate (IDEAM) tags}
\usage{
get_climate_tags(language = "ES")
}
\arguments{
\item{language}{character with the language of the tags (\code{"EN"} or
\code{"ES"}. Default is \code{"ES"}.}
}
\value{
\code{data.frame} object with available climate tags.
}
\description{
Retrieve available climate tags to be consulted. The list is only available
in Spanish.
}
\examples{
dict <- get_climate_tags("ES")
head(dict)

}
