% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgn_cnpv.R
\name{get_mgn_cnpv_mun}
\alias{get_mgn_cnpv_mun}
\title{Retrieve data from the 2018 National Population and Living Census}
\usage{
get_mgn_cnpv_mun(mun = "all", include_geometry = TRUE, columns = "all")
}
\arguments{
\item{mun}{character vector containing the DIVIPOLA official codes or the
municipalities names. Default is "all"}

\item{include_geometry}{boolean stating either to include or not the spatial
geometry}

\item{columns}{character vector indicating which columns are meant to be
included in the consultation}
}
\value{
a data.frame containing the consulted municipalities and desired
columns
}
\description{
Retrieve data from the 2018 National Population and Living Census
}
\examples{
\dontrun{
get_mgn_cnpv_mun()
}

}
