% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_IDEAM.R
\name{group_stations}
\alias{group_stations}
\title{Group stations}
\usage{
group_stations(stations_df, tag)
}
\arguments{
\item{stations_df}{data.frame containing stations to group}

\item{tag}{string with the tag consulted}
}
\value{
a data.frame with only two columns containing the dates and the
grouped observations
}
\description{
Groups stations by the mean of the observations
}
\keyword{internal}
