% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{list_data}
\alias{list_data}
\title{Download list of available datasets}
\usage{
list_data(group = "all")
}
\arguments{
\item{group}{Specific group to be consulted ("demographic", "geospatial",
"climate")}
}
\value{
tibble containing listed available datasets
}
\description{
Download list of available datasets
}
\examples{
list_data("geospatial")

}
