% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{list_datasets}
\alias{list_datasets}
\title{Download list of available datasets}
\usage{
list_datasets(module = "all")
}
\arguments{
\item{module}{Specific group to be consulted ("demographic", "geospatial",
"climate")}
}
\value{
tibble containing listed available datasets
}
\description{
Download list of available datasets
}
\examples{
list_datasets("geospatial")

}
