% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{look_up}
\alias{look_up}
\title{Filter list of available datasets based on keywords given by the user}
\usage{
look_up(module = "all", keywords, logic = "or")
}
\arguments{
\item{module}{character with module to be consulted (\code{"demographic"},
\code{"geospatial"}, \code{"climate"})}

\item{keywords}{character or vector of characters to be look up in the
description}

\item{logic}{A character string specifying the matching logic.
Can be either "or" or "and".
\code{logic = "or"}: Matches rows containing at least one of the specified
keywords in their descriptions.
\code{logic = "and"}: Matches rows containing all of the specified keywords
in their descriptions.}
}
\value{
\code{data.frame} object with the available datasets
}
\description{
Filter list of available datasets based on keywords given by the user
}
\section{List available datasets containing user-specified keywords in their}{
descriptions.
}

\examples{
filtered_datasets <- look_up("demographic", c("sex", "age"), "and")

}
