% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_geo_dem.R
\name{merge_geo_dem}
\alias{merge_geo_dem}
\title{Match and merge geospatial and demographic datasets}
\usage{
merge_geo_dem(spatial_level, dem_dataset, column)
}
\arguments{
\item{spatial_level}{character with the level of spatial aggregation to be used.
\code{"department"} or \code{"municipality"}}

\item{dem_dataset}{character with the demographic dataset name}

\item{column}{character with the column we are interested in. It depends on
the demographic dataset, so it is important to have checked it before}
}
\value{
\code{data.frame} object with the merged data
}
\description{
This function downloads a demographic and a geospatial dataset, does a
pivot to the demographic dataset based on a column of interest and merges
with a geospatial dataset (with \code{include_cnpv} as \code{FALSE}.
Since the smallest level of spatial aggregation present in the demographic
datasets is municipality, this functions can only match with the municipality
and department code of geospatial datasets.
}
\examples{
dem_file <- "DANE_CNPVV_2018_9VD"
merged <- merge_geo_dem("department", dem_file, "tipo_de_servicio_sanitario")

}
