% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_geo_demographic.R
\name{merge_geo_demographic}
\alias{merge_geo_demographic}
\title{Match and merge geospatial and demographic datasets}
\usage{
merge_geo_demographic(demographic_dataset)
}
\arguments{
\item{demographic_dataset}{character with the demographic dataset name.
Please use \code{list_datasets("demographic")} to check available datasets}
}
\value{
\code{data.frame} object with the merged data
}
\description{
This function adds the key information of a demographic dataset to a
geospatial dataset based on the spatial aggregation level. Since the smallest
level of spatial aggregation present in the demographic datasets is
municipality, this function can only merge with geospatial datasets that
present municipality or department level.
}
\examples{
merged <- merge_geo_demographic("DANE_CNPVV_2018_9VD")
head(merged)
}
