% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{monthly_bshg}
\alias{monthly_bshg}
\title{Calculate monthly sunshine duration}
\usage{
monthly_bshg(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the month}
}
\value{
numeric value calculated
}
\description{
Calculate monthly sunshine duration
}
\keyword{internal}
