% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{monthly_ptpm}
\alias{monthly_ptpm}
\title{Calculate monthly precipitation}
\usage{
monthly_ptpm(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the month}
}
\value{
numeric value calculated
}
\description{
Calculate monthly precipitation
}
\keyword{internal}
