% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{monthly_tmn}
\alias{monthly_tmn}
\title{Calculate monthly minimum temperature}
\usage{
monthly_tmn(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the month}
}
\value{
numeric value calculated
}
\description{
Calculate monthly minimum temperature
}
\keyword{internal}
