% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{monthly_tmx}
\alias{monthly_tmx}
\title{Calculate monthly maximum temperature}
\usage{
monthly_tmx(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the month}
}
\value{
numeric value calculated
}
\description{
Calculate monthly maximum temperature
}
\keyword{internal}
