% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_aggregation_rules.R
\name{monthly_tssm}
\alias{monthly_tssm}
\title{Calculate monthly dry-bulb mean temperature}
\usage{
monthly_tssm(group)
}
\description{
Calculate monthly dry-bulb mean temperature
}
\keyword{internal}
