% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{name_to_code_dep}
\alias{name_to_code_dep}
\title{Retrieve departments' DIVIPOLA codes from names}
\usage{
name_to_code_dep(department_name)
}
\arguments{
\item{department_name}{character vector with the names of the departments.}
}
\value{
character vector with the DIVIPOLA codes of the departments.
}
\description{
Retrieve departments' DIVIPOLA codes from their names.
}
\examples{
dptos <- c("Tolima", "Huila", "Amazonas")
name_to_code_dep(dptos)

}
