% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_weather.R
\name{plot_stations}
\alias{plot_stations}
\title{Plot stations}
\usage{
plot_stations(stations_df, tag)
}
\arguments{
\item{stations_df}{data.frame with stations containing observations for each
date}

\item{tag}{string with the tag consulted}
}
\description{
Plot weather stations if required
}
\keyword{internal}
