% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_climate}
\alias{retrieve_climate}
\title{Retrieve climate (.data) file from one station}
\usage{
retrieve_climate(dataset_path, start_date, end_date)
}
\arguments{
\item{dataset_path}{character path to the climate dataset on server}

\item{start_date}{character with the first date to consult in the format
\code{"YYYY-MM-DD"}}

\item{end_date}{character with the last date to consult in the format
\code{"YYYY-MM-DD"} (Last available date is \code{"2023-05-31"})}
}
\value{
\code{data.frame} object with downloaded data filtered for requested
dates
}
\description{
Retrieve climate (.data) file from one station
}
\keyword{internal}
