% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_csv}
\alias{retrieve_csv}
\title{Retrieve csv file}
\usage{
retrieve_csv(dataset_path, new_dir_path, new_dir, sep)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}

\item{new_dir_path}{path to the created directory}

\item{new_dir}{directory to store data}

\item{sep}{separator for csv data}
}
\value{
dataset
}
\description{
Retrieve csv file
}
\keyword{internal}
