% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieve data file}
\usage{
retrieve_data(dataset_path)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}
}
\value{
dataset
}
\description{
Retrieve data file
}
\keyword{internal}
