% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_dataset}
\alias{retrieve_dataset}
\title{Retrieve dataset from path}
\usage{
retrieve_dataset(dataset_path)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}
}
\value{
consulted dataset
}
\description{
Retrieve dataset from path
}
\keyword{internal}
