% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_demographic.R
\name{retrieve_demographic_dataset}
\alias{retrieve_demographic_dataset}
\title{Retrieve demographic dataset from path}
\usage{
retrieve_demographic_dataset(dataset_path, sheet)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}

\item{sheet}{string to indicate specific sheet}
}
\value{
consulted dataset
}
\description{
Retrieve demographic dataset from path
}
\keyword{internal}
