% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_dict_path}
\alias{retrieve_dict_path}
\title{Retrieve dictionary path of named dataset}
\usage{
retrieve_dict_path(dict_name)
}
\arguments{
<<<<<<< Updated upstream
\item{dict_name}{character with the dictionary name.}
=======
\item{dict_name}{character with the dictionary file name.}
>>>>>>> Stashed changes
}
\value{
character with path to retrieve the dataset.
}
\description{
Dictionaries are not included in the general documentation file. Therefore,
the path is built internally.
}
\keyword{internal}
