% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{retrieve_dictionary}
\alias{retrieve_dictionary}
\title{Retrieve dictionary from path}
\usage{
retrieve_dictionary(dataset_path)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}
}
\value{
consulted dictionary.
}
\description{
Retrieve dictionary from path
}
\keyword{internal}
