% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_excel}
\alias{retrieve_excel}
\title{Retrieve excel file}
\usage{
retrieve_excel(dataset_path, new_dir_path, new_dir, sheet)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}

\item{new_dir_path}{path to the created directory}

\item{new_dir}{directory to store data}

\item{sheet}{string to indicate specific sheet}
}
\value{
dataset
}
\description{
Retrieve excel file
}
\keyword{internal}
