% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geospatial.R
\name{retrieve_geospatial_dataset}
\alias{retrieve_geospatial_dataset}
\title{Retrieve geospatial dataset from path}
\usage{
retrieve_geospatial_dataset(dataset_path)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}

\item{sheet}{string to indicate specific sheet}
}
\value{
consulted dataset
}
\description{
Retrieve geospatial dataset from path
}
\keyword{internal}
