% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geospatial.R
\name{retrieve_geospatial_name}
\alias{retrieve_geospatial_name}
\title{Retrieve geospatial dataset name for consultation}
\usage{
retrieve_geospatial_name(spatial_level)
}
\arguments{
\item{spatial_level}{character with the spatial level to be consulted.}
}
\value{
character containing the geospatial dataset name. If the input is
invalid an error will be thrown.
}
\description{
Retrieve a geospatial dataset name from the spatial level. Checks the
existence of the spatial level and datasets.
}
\keyword{internal}
