% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_path}
\alias{retrieve_path}
\title{Retrieve path of named dataset}
\usage{
retrieve_path(dataset)
}
\arguments{
\item{dataset}{name of the consulted dataset}
}
\value{
path to retrieve the dataset from server
}
\description{
Retrieve path of named dataset
}
\keyword{internal}
