% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{retrieve_stations_data}
\alias{retrieve_stations_data}
\title{Retrieve data from named stations for a specific tag}
\usage{
retrieve_stations_data(
  stations,
  start_date,
  end_date,
  frequency,
  tag,
  aggregate = TRUE
)
}
\arguments{
\item{stations}{numeric vector containing the stations' codes}

\item{start_date}{character with the first date to consult in the format
\code{"YYYY-MM-DD"}}

\item{end_date}{character with the last date to consult in the format
\code{"YYYY-MM-DD"} (Last available date is \code{"2023-05-31"})}

\item{frequency}{character with the aggregation frequency. (\code{"day"},
\code{"week"}, \code{"month"} or \code{"year"})}

\item{tag}{unique character containing climate tags to consult}

\item{aggregate}{logical for data aggregation, if \code{FALSE}, returns the
data from each individual station in the area. (Default = \code{TRUE})}
}
\value{
\code{data.frame} with observations from the requested stations
}
\description{
Retrieve climate data from a list of stations under the same tag.This data is
retrieved from local meteorological stations provided by IDEAM
}
\keyword{internal}
