% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_support_path}
\alias{retrieve_support_path}
\title{Retrieve support dataset path}
\usage{
retrieve_support_path(dataset)
}
\arguments{
\item{dataset}{character with the support dataset name}
}
\value{
character with path to retrieve the dataset from server
}
\description{
Support data is used for internal purposes and they are not included in the
general documentation file
}
\keyword{internal}
