% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_value_key}
\alias{retrieve_value_key}
\title{Retrieve value from key}
\usage{
retrieve_value_key(key)
}
\arguments{
\item{key}{character key}
}
\value{
character containing value
}
\description{
Retrieve value from key included in configuration file
}
\keyword{internal}
