% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_weather.R
\name{retrieve_working_stations}
\alias{retrieve_working_stations}
\title{Retrieve stations that took measurements in the given dates}
\usage{
retrieve_working_stations(
  stations,
  start_date,
  end_date,
  frequency,
  tag,
  plot = FALSE,
  group = FALSE
)
}
\arguments{
\item{stations}{vector containing the stations codes}

\item{start_date}{starting date in format "YYYY-MM-DD"}

\item{end_date}{end date in format "YYYY-MM-DD"}

\item{frequency}{aggregation frequency. Can be "day", "week","month" or
"year"}

\item{tag}{character containing tag to analyze}

\item{plot}{if TRUE, plot the individual stations that contain data}

\item{group}{if TRUE, returns only one observation from the mean of the}
}
\value{
data.frame containing the observed data for the given stations
}
\description{
Retrieves, from a list of stations, the ones that contain observations for
the given tag in specified dates. Might include stations that are no longer
working.
}
\keyword{internal}
