% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_zip}
\alias{retrieve_zip}
\title{Retrieve zip file}
\usage{
retrieve_zip(dataset_path, new_dir_path, new_dir)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}

\item{new_dir_path}{path to the created directory}

\item{new_dir}{directory to store data}
}
\value{
dataset
}
\description{
Retrieve zip file
}
\keyword{internal}
