% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{summarise}
\alias{summarise}
\title{Summary and group climate data according to time frequency}
\usage{
summarise(.data, tag, frequency)
}
\arguments{
\item{.data}{\code{data.frame} containing at least one observed column}

\item{tag}{unique character with the tag consulted}
}
\value{
\code{data.frame} with grouped observations
}
\description{
Summary and group climate data according to time frequency
}
\keyword{internal}
