% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{trans_divipola_department}
\alias{trans_divipola_department}
\title{Translate department names to  official municipalities' DIVIPOLA names}
\usage{
trans_divipola_department(department_name)
}
\arguments{
\item{department_name}{character vector with the names to be translated}
}
\value{
character vector with the DIVIPOLA name of the departments
}
\description{
Translate department names to their respective official names from DIVIPOLA
}
\examples{
dptos <- c("Bogota DC", "San Andres")
official_names <- trans_divipola_department(dptos)

}
