% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divipola.R
\name{trans_divipola_municipality}
\alias{trans_divipola_municipality}
\title{Translate municipality names to  official municipalities' DIVIPOLA names}
\usage{
trans_divipola_municipality(department_name, municipality_name)
}
\arguments{
\item{department_name}{character vector with the names of the
departments containing the municipalities}

\item{municipality_name}{character vector with the names to be translated}
}
\value{
character vector with the DIVIPOLA name of the municipalities
}
\description{
Translate municipalities' names to their respective official names from
DIVIPOLA
}
\examples{

dptos <- c("Bogota", "Antioquia", "Tolima")
mpios <- c("Bogota DC", "Carmen dell Viboral", "CarmendeApicala")
official_names <- trans_divipola_municipality(dptos, mpios)

}
