% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_weather.R
\name{weather_data}
\alias{weather_data}
\title{Retrieve weather data from stations in geometry}
\usage{
weather_data(
  geometry,
  start_date,
  end_date,
  frequency,
  tags,
  plot = FALSE,
  group = FALSE
)
}
\arguments{
\item{geometry}{sf object containing a geometry (polygon or multipolygon)}

\item{start_date}{starting date in format "YYYY-MM-DD"}

\item{end_date}{end date in format "YYYY-MM-DD"}

\item{frequency}{aggregation frequency. Can be "day", "week","month" or
"year"}

\item{tags}{character containing tags to analyze}

\item{plot}{if TRUE, plot the individual stations that contain data}

\item{group}{if TRUE, returns only one observation from the mean of the
stations consulted}
}
\value{
data.frame with the observed data for the given geometry
}
\description{
Retrieve weather stations with observed data in a given geometry.
}
\examples{
\dontrun{
weather_data(geometry, "2021-11-14", "2021-11-30", "day", "TSSM_CON")
}

}
