% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_IDEAM.R
\name{weather_stations}
\alias{weather_stations}
\title{Retrieve weather stations}
\usage{
weather_stations(
  geometry,
  start_date,
  end_date,
  frequency,
  tag,
  plot = FALSE,
  group = FALSE
)
}
\arguments{
\item{geometry}{sf object containing a geometry (polygon or multipolygon)}

\item{start_date}{starting date in format "YYYY-MM-DD"}

\item{end_date}{end date in format "YYYY-MM-DD"}

\item{frequency}{aggregation frequency. Can be "day", "week","month" or "year"}

\item{tag}{character containing tags to analyze}

\item{plot}{if TRUE, plot the individual stations that contain data}

\item{group}{if TRUE, returns only one observation from the mean of the
stations consulted}
}
\value{
data.frame with the observed data for the given geometry
}
\description{
Retrieve weather stations with observed data in a given geometry.
}
