% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate climate data according to time frequency}
\usage{
aggregate(.data, tag, frequency)
}
\arguments{
\item{.data}{data.frame containing at least one observed column}

\item{tag}{string with the tag consulted}
}
\value{
a data.frame with aggregated observations
}
\description{
Aggregate climate data according to time frequency
}
\keyword{internal}
