% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate stations' data}
\usage{
aggregate(stations_data, tag)
}
\arguments{
\item{stations_data}{data.frame containing stations' data to aggregate}

\item{tag}{character with the consulted tag}
}
\value{
\code{data.frame} with only two columns containing the dates and the
aggregated observations
}
\description{
Aggregate stations' data
}
\keyword{internal}
