% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{aggregate_climate}
\alias{aggregate_climate}
\title{Aggregate climate data for different frequencies}
\usage{
aggregate_climate(.data, frequency)
}
\arguments{
\item{.data}{\code{data.frame} obtained from download functions. Can only
include observations under the same tag.}

\item{frequency}{character with the aggregation frequency. (\code{"day"},
\code{"month"} or \code{"year"})}
}
\value{
\code{data.frame} with the aggregated data for specific frequency
}
\description{
Aggregate time series downloaded data from specific tags to day, month or
year. Only observations under the tags \code{TSSM_CON}, \code{TMN_CON},
\code{TMX_CON}, \code{PTPM_CON}, and \code{BSHG_CON} can be aggregated,
since are the ones where methodology  for aggregation is explicitly available
from the source.
}
\examples{
\dontrun{
aggregate_tssm(tssm, "day")
}

}
