% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_aggregation_rules.R
\name{annual_tssm}
\alias{annual_tssm}
\title{Calculate annual dry-bulb mean temperature}
\usage{
annual_tssm(group)
}
\description{
Calculate annual dry-bulb mean temperature
}
\keyword{internal}
