% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{climate_stations}
\alias{climate_stations}
\title{Retrieve climate from stations}
\usage{
climate_stations(
  stations,
  start_date,
  end_date,
  frequency,
  tags,
  group = FALSE
)
}
\arguments{
\item{stations}{Vector containing the stations codes.}

\item{start_date}{Character with the first date to consult in format
"YYYY-MM_DD".}

\item{end_date}{Character with the last date to consult in format
"YYYY-MM_DD".}

\item{frequency}{Character with the aggregation frequency. Can be
"day", "week","month" or"year".}

\item{tags}{Character containing tags to consult.}

\item{group}{Boolean for data aggregation, if FALSE, returns the data from
each individual station in ROI.(Default = FALSE).}
}
\value{
data.frame with the observed data for the given geometry
}
\description{
Retrieve climate stations with observed data in a given geometry.
}
\examples{
\dontrun{
climate_stations(stations, "2021-11-14", "2021-11-30", "day", "TSSM_CON")
}

}
