% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_tags.R
\docType{data}
\name{climate_tags}
\alias{climate_tags}
\title{climate_tags}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
data(climate_tags)
}
\description{
dictionary for climate tags
}
\details{
Dictionary for climate tags
}
\keyword{datasets}
