% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_aggregation_rules.R
\name{daily_tssm}
\alias{daily_tssm}
\title{Climate aggregation rules}
\usage{
daily_tssm(group)
}
\arguments{
\item{group}{\code{data.frame} object with filtered and grouped data.}
}
\value{
numeric value calculated.
}
\description{
Climate temporal aggregation rules are provided by the source, and guarantee
data quality given missing information. These rules are included in the
package to make the download and aggregation process easier for the user. The
aggregation is not available for all climate data, and is only available for
information under the tags \code{TSSM_CON}, \code{TMN_CON}, \code{TMX_CON},
\code{PTPM_CON}, and \code{BSHG_CON}. Internal functions are provided as a
set of comprehensible rules to aggregate the data for daily, monthly and
annual frequencies.
}
\section{Methods}{

Aggregation can only be performed from the previous level, meaning for
monthly aggregation, the data must be already aggregated daily, and for
annual aggregation the data must be monthly.
}

\keyword{internal}
