% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_list.R
\docType{data}
\name{datasets_list}
\alias{datasets_list}
\title{datasets_list}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(datasets_list)
}
\description{
list of datasets description in spanish and english
}
\details{
List containing both datasets description in spanish and english
}
\keyword{datasets}
