% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{dictionary}
\alias{dictionary}
\title{Download data dictionary for geospatial data}
\usage{
dictionary(dataset)
}
\arguments{
\item{dataset}{character with the dataset name}
}
\value{
\code{data.frame} with data dictionary
}
\description{
Download data dictionary for geospatial data
}
\examples{
dictionary("DANE_MGN_2018_SETU")

}
