% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{dictionary}
\alias{dictionary}
\title{Download data dictionaries}
\usage{
dictionary(dataset)
}
\arguments{
\item{dataset}{character with the dataset name}
}
\value{
\code{data.frame} object with data dictionary
}
\description{
Demographic and climate datasets contain data dictionaries to understand
internal tags and named columns
}
\examples{
dict <- dictionary("DANE_MGN_2018_SETU")

}
