% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{dictionary}
\alias{dictionary}
\title{Download data dictionary for geospatial dataframes}
\usage{
dictionary(dataset)
}
\arguments{
\item{dataset}{String indicating dataset code.}
}
\value{
tibble containing data dictionary
}
\description{
Download data dictionary for geospatial dataframes
}
\examples{
data_dictionary("DANE_MGNCNPV_2018_SETU")

}
