% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate_geom}
\alias{download_climate_geom}
\title{Retrieve climate data from stations in geometry}
\usage{
download_climate_geom(
  geometry,
  start_date,
  end_date,
  frequency,
  tags,
  aggregate = FALSE
)
}
\arguments{
\item{geometry}{sf object containing the geometry for a given ROI. This
geometry can be either a POLYGON or MULTIPOLYGON.}

\item{start_date}{Character with the first date to consult in format
"YYYY-MM-DD".}

\item{end_date}{Character with the last date to consult in format
"YYYY-MM-DD". Last available date is 2023/05/31.}

\item{frequency}{Character with the aggregation frequency. Can be
"day", "week","month" or "year".}

\item{tags}{Character containing tags to consult.}

\item{aggregate}{Boolean for data aggregation, if FALSE, returns the data from
each individual station in ROI.(Default = FALSE).}
}
\value{
data.frame with the observed data for the given geometry.
}
\description{
Retrieve climate stations with observed data in a given geometry. This
data is retrieved from local meteorological stations.
}
\examples{
\dontrun{
download_climate_geom(geometry, "2021-11-14", "2021-11-30", "day", "TSSM_CON")
}

}
