% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{download_climate_stations}
\alias{download_climate_stations}
\title{Download climate data from stations}
\usage{
download_climate_stations(stations, start_date, end_date, tags)
}
\arguments{
\item{stations}{\code{data.frame} containing the stations' codes. This
\code{data.frame} must be retrieved from the function
\code{stations_in_roi()}}

\item{start_date}{character with the first date to consult in the format
\code{"YYYY-MM-DD"}}

\item{end_date}{character with the last date to consult in the format
\code{"YYYY-MM-DD"} (Last available date is \code{"2023-05-31"})}

\item{tags}{character containing climate tags to consult}
}
\value{
\code{data.frame} with observations from the requested stations
}
\description{
Download climate data from named stations.This data is
retrieved from local meteorological stations provided by IDEAM.
}
\examples{
\dontrun{
tssm <- download_climate_stations(
  stations, "2021-11-14", "2021-11-20", "TSSM_CON"
)
print(tssm)
}

}
