% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_demographic.R
\name{download_demographic}
\alias{download_demographic}
\title{Download demographic dataset}
\usage{
download_demographic(dataset, sheet = "None")
}
\arguments{
\item{dataset}{dataset code}

\item{sheet}{string to indicate specific sheet (internal Dataset)}
}
\value{
dataframe downloaded
}
\description{
Download demographic named dataset from server. This data is obtained from
the national census of 2018.
}
\examples{
download_demographic("DANE_CNPV_2018_Hogares", "1HD")

}
