% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_demographic.R
\name{download_demographic}
\alias{download_demographic}
\title{Download demographic dataset}
\usage{
download_demographic(dataset, sheet = "None")
}
\arguments{
\item{dataset}{String indicating dataset code.}

\item{sheet}{String indicating specific sheet (internal dataset).}
}
\value{
data.frame downloaded.
}
\description{
Download demographic dataset. This data is obtained from the national census
of 2018.
}
\examples{
download_demographic("DANE_CNPV_2018_Hogares", "1HD")

}
