% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_demographic.R
\name{download_demographic}
\alias{download_demographic}
\title{Download demographic dataset}
\usage{
download_demographic(dataset)
}
\arguments{
\item{dataset}{character with the dataset name}
}
\value{
\code{data.frame} with downloaded data.
}
\description{
This function downloads demographic datasets from the National Population
and Housing Census (CNPV)
}
\examples{
dem <- download_demographic("DANE_CNPVH_2018_1HD")
print(dem)

}
