% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_demographic}
\alias{download_demographic}
\title{Download dataset}
\usage{
download_demographic(dataset, sheet = "None")
}
\arguments{
\item{dataset}{dataset code}

\item{sheet}{string to indicate specific sheet}
}
\value{
dataframe downloaded
}
\description{
Download dataset
}
\examples{
download_demographic("DANE_MGNCNPV_2018_DPTO")

}
