% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_demographic.R
\name{download_demographic}
\alias{download_demographic}
\title{Download demographic dataset}
\usage{
download_demographic(dataset)
}
\arguments{
\item{dataset}{String indicating dataset code.}
}
\value{
data.frame downloaded.
}
\description{
Download demographic dataset. This data is obtained from the national census
of 2018.
}
\examples{
download_demographic("DANE_CNPVH_2018_1HD")

}
