% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geospatial.R
\name{download_geospatial}
\alias{download_geospatial}
\title{Download geospatial dataset}
\usage{
download_geospatial(dataset, geospatial = T, demographic = T)
}
\arguments{
\item{dataset}{character with the dataset name}

\item{geospatial}{boolean for presenting or not geospatial related columns}

\item{demographic}{boolean for presenting or not demographic related columns}
}
\value{
\code{sf} \code{data.frame} object with structures' details
}
\description{
This function downloads geospatial datasets from the Geostatistical National
Framework (MGN) at different levels of aggregation. These datasets include a
summarized version of the National Population and Dwelling Census (CNPV)
}
\examples{
dptos <- download_geospatial("DANE_MGN_2018_DPTO", T, F)
print(dptos)

}
