% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geospatial.R
\name{download_geospatial}
\alias{download_geospatial}
\title{Download geospatial dataset}
\usage{
download_geospatial(dataset)
}
\arguments{
\item{dataset}{dataset code}
}
\value{
data.frame downloaded
}
\description{
Download geospatial named dataset from server. This data is obtained from
the national census of 2018.
}
\examples{
download_geospatial("DANE_MGNCNPV_2018_DPTO")

}
