% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geospatial.R
\name{download_geospatial}
\alias{download_geospatial}
\title{Download geospatial dataset}
\usage{
download_geospatial(dataset)
}
\arguments{
\item{dataset}{character with the dataset name}
}
\value{
\code{sf} \code{data.frame} object with structures' details and geometries
}
\description{
This function downloads geospatial datasets from the Geostatistical National
Framework (MGN) at different levels of aggregation. These datasets include a
summarized version of the National Population and Housing Census (CNPV)
}
\examples{
dptos <- download_geospatial("DANE_MGNCNPV_2018_DPTO")
print(dptos)

}
