% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geospatial.R
\name{download_geospatial}
\alias{download_geospatial}
\title{Download geospatial dataset}
\usage{
download_geospatial(dataset, include_geom = TRUE, include_cnpv = TRUE)
}
\arguments{
\item{dataset}{character with the geospatial dataset name}

\item{include_geom}{logical for including (or not) the geometry.
Default is \code{TRUE}}

\item{include_cnpv}{logical for including (or not) CNPV demographic and
socioeconomic information. Default is \code{TRUE}}
}
\value{
\code{data.frame} object with downloaded data
}
\description{
This function downloads geospatial datasets from the Geostatistical National
Framework (MGN) at different levels of spatial aggregation. These datasets
include a summarized version of the National Population and Dwelling Census
(CNPV) with demographic and socioeconomic information.
}
\examples{
departments <- download_geospatial("DANE_MGN_2018_DPTO", TRUE, FALSE)

}
