% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geospatial.R
\name{download_geospatial}
\alias{download_geospatial}
\title{Download geospatial dataset}
\usage{
download_geospatial(dataset)
}
\arguments{
\item{dataset}{String indicating dataset code.}
}
\value{
data.frame downloaded data.
}
\description{
Download geospatial named dataset from server. This data is merged with a
simplified version of the National Census, associating the census with each
geospatial structure.
}
\examples{
download_geospatial("DANE_MGNCNPV_2018_DPTO")

}
