% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_mgn_cnpv_ad1}
\alias{filter_mgn_cnpv_ad1}
\title{Filter department dataset}
\usage{
filter_mgn_cnpv_ad1(census_data, codes, columns)
}
\arguments{
\item{census_data}{MGN - CNPV dataset}

\item{codes}{Administrative division to include}

\item{columns}{Columns to include}
}
\value{
filtered dataset for department level
}
\description{
Filter department dataset
}
\keyword{internal}
