% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{list_datasets}
\alias{list_datasets}
\title{Download list of available datasets}
\usage{
list_datasets(module = "all")
}
\arguments{
\item{module}{character with module to be consulted (\code{"demographic"},
\code{"geospatial"}, \code{"climate"})}
}
\value{
\code{data.frame} object with the available datasets
}
\description{
List all available datasets by name, including category, description and
source
}
\examples{
list <- list_datasets("geospatial")

}
