% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_geo_demographic.R
\name{merge_geo_demographic}
\alias{merge_geo_demographic}
\title{Match and merge geospatial and demographic datasets}
\usage{
merge_geo_demographic(demographic_dataset, column)
}
\arguments{
\item{demographic_dataset}{character with the demographic dataset name.
Please use \code{list_datasets("demographic")} to check available datasets}

\item{column}{character with the column we are interested in. It depends on
the demographic dataset, so it is important to have checked the dataset
beforehand}
}
\value{
\code{data.frame} object with the merged data
}
\description{
This function downloads a demographic and a geospatial dataset, does a
pivot to the demographic dataset based on a column of interest and merges
with a geospatial dataset according to the level of the demographic dataset.
Since the smallest level of spatial aggregation present in the demographic
datasets is municipality, this functions can only match with the municipality
and department code of geospatial datasets.
}
\examples{
census_file <- "DANE_CNPVV_2018_9VD"
merged <- merge_geo_demographic(census_file, "tipo_de_servicio_sanitario")

}
