% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_climate.R
\name{monthly_tssm}
\alias{monthly_tssm}
\title{Calculate monthly dry-bulb mean temperature}
\usage{
monthly_tssm(group)
}
\arguments{
\item{group}{\code{data.frame} with filtered and grouped data for the month}
}
\value{
numeric value calculated
}
\description{
Calculate monthly dry-bulb mean temperature
}
\keyword{internal}
