% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_climate_path}
\alias{retrieve_climate_path}
\title{Retrieve climate directory path}
\usage{
retrieve_climate_path()
}
\value{
character with path to retrieve the dataset from server
}
\description{
Climate data is retrieved from a general directory. Path is build for said
directory
}
\keyword{internal}
