% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_dict_path}
\alias{retrieve_dict_path}
\title{Retrieve dictionary path of named dataset}
\usage{
retrieve_dict_path(dataset)
}
\arguments{
\item{dataset}{character with the dictionary name}
}
\value{
character with path to retrieve the dataset from server
}
\description{
Dictionaries are not included in the general documentation file. Therefore,
the path is built internally
}
\keyword{internal}
