% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_path}
\alias{retrieve_path}
\title{Retrieve path of named dataset}
\usage{
retrieve_path(dataset)
}
\arguments{
\item{dataset}{character with the dataset name}
}
\value{
character with path to retrieve the dataset from server
}
\description{
Datasets are included in the general documentation file. Only dictionaries
and support files are treated differently and not included in the file.
}
\keyword{internal}
