% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{retrieve_stations_data}
\alias{retrieve_stations_data}
\title{Retrieve data from stations that took measurements in the given dates}
\usage{
retrieve_stations_data(
  stations,
  start_date,
  end_date,
  frequency,
  tag,
  aggregate = TRUE
)
}
\arguments{
\item{stations}{Vector containing the stations' codes.}

\item{start_date}{Character with the first date to consult in format
"YYYY-MM-DD".}

\item{end_date}{Character with the last date to consult in format
"YYYY-MM-DD". Last available date is 2023/05/31.}

\item{frequency}{Character with the aggregation frequency. Can be
"day", "week","month" or "year".}

\item{tag}{Character containing tag to consult.}

\item{aggregate}{Boolean for data aggregation, if FALSE, returns the data from
each individual station in ROI.(Default = TRUE).}
}
\value{
data.frame containing the observed data for the given stations.
}
\description{
Retrieves, from a list of stations, the ones that contain observations for
the given tag in specified dates. Might include stations that are no longer
working.
}
\keyword{internal}
