% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_table}
\alias{retrieve_table}
\title{Retrieve csv file}
\usage{
retrieve_table(dataset_path, sep)
}
\arguments{
\item{dataset_path}{path to the dataset on repository}

\item{sep}{separator for csv data}
}
\value{
dataset
}
\description{
Retrieve csv file
}
\keyword{internal}
