% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_table}
\alias{retrieve_table}
\title{Retrieve table (csv and data) file}
\usage{
retrieve_table(dataset_path, sep = ";")
}
\arguments{
\item{dataset_path}{character path to the dataset on server.}

\item{sep}{separator for table data.}
}
\value{
\code{data.frame} object with downloaded data.
}
\description{
Retrieve table (csv and data) file
}
\keyword{internal}
