% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_zip}
\alias{retrieve_zip}
\title{Retrieve zip file}
\usage{
retrieve_zip(dataset_path, dataset)
}
\arguments{
\item{dataset_path}{Path to dataset on repository}

\item{dataset}{Dataset name to download}
}
\value{
dataset
}
\description{
Retrieve zip file
}
\keyword{internal}
