% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{retrieve_zip}
\alias{retrieve_zip}
\title{Retrieve zip file}
\usage{
retrieve_zip(dataset_path, dataset_name)
}
\arguments{
\item{dataset_path}{character with the path to dataset on repository}

\item{dataset_name}{character with the dataset name}
}
\value{
\code{sf} data.frame object with structures' details and geometries
}
\description{
Retrieve zip file
}
\keyword{internal}
