% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{set_magna_sirgas}
\alias{set_magna_sirgas}
\title{Change coordinate system to Magna Sirgas}
\usage{
set_magna_sirgas(.data)
}
\arguments{
\item{.data}{sf object containing coordinates and established geometry}
}
\value{
sf dataframe with Colombian coordinate reference system
}
\description{
Allows the user to set the Coordinate Reference System to local
(Magna Sirgas)
}
\examples{
\dontrun{
set_magna_sirgas(MGNCNPV_MUN_2018)
}

}
