% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_weather.R
\name{stations_in_roi}
\alias{stations_in_roi}
\title{Stations in region of interest}
\usage{
stations_in_roi(geometry)
}
\arguments{
\item{geometry}{sf object containing a geometry (polygon or multipolygon)}
}
\value{
data.frame with the stations inside the consulted geometry
}
\description{
Retrieves the stations contained inside a given geometry
}
\examples{
\dontrun{
stations_in_roy(geometry)
}

}
