% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_climate.R
\name{summarise}
\alias{summarise}
\title{Summary and group of climate data according to time frequency}
\usage{
summarise(.data, tag, frequency)
}
\arguments{
\item{.data}{data.frame containing at least one observed column}

\item{tag}{string with the tag consulted}
}
\value{
a data.frame with aggregated observations
}
\description{
Summary and group of climate data according to time frequency
}
\keyword{internal}
